# docs/openapi.json
{
  "components": {
    "schemas": {
...
      "Body_enroll_face_api_personas__person_id__faces_post": {
        "properties": {
          "image": {
            "contentMediaType": "application/octet-stream",
            "description": "Foto con exactamente un rostro (JPEG, PNG, WebP o BMP)",
            "title": "Image",
            "type": "string"
          }
        },
        "required": [
          "image"
        ],
        "title": "Body_enroll_face_api_personas__person_id__faces_post",
        "type": "object"
      },
...
      "ErrorOut": {
        "properties": {
          "code": {
            "title": "Code",
            "type": "string"
          },
          "detail": {
            "title": "Detail",
            "type": "string"
          }
        },
        "required": [
          "detail",
          "code"
        ],
        "title": "ErrorOut",
        "type": "object"
      },
...
      "FaceEnrollmentOut": {
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "face_count": {
            "title": "Face Count",
            "type": "integer"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "model_version": {
            "title": "Model Version",
            "type": "string"
          },
          "person_id": {
            "format": "uuid",
            "title": "Person Id",
            "type": "string"
          },
          "quality": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/FaceQualityOut"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "required": [
          "id",
          "model_version",
          "created_at",
          "person_id",
          "quality",
          "face_count"
        ],
        "title": "FaceEnrollmentOut",
        "type": "object"
      },
...
      "FaceQualityOut": {
        "properties": {
          "brightness": {
            "title": "Brightness",
            "type": "number"
          },
          "detection_score": {
            "title": "Detection Score",
            "type": "number"
          },
          "face_size": {
            "title": "Face Size",
            "type": "integer"
          },
          "sharpness": {
            "title": "Sharpness",
            "type": "number"
          }
        },
        "required": [
          "face_size",
          "detection_score",
          "sharpness",
          "brightness"
        ],
        "title": "FaceQualityOut",
        "type": "object"
      },
...
      "FaceSampleOut": {
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "model_version": {
            "title": "Model Version",
            "type": "string"
          }
        },
        "required": [
          "id",
          "model_version",
          "created_at"
        ],
        "title": "FaceSampleOut",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "PersonCreate": {
        "additionalProperties": false,
        "properties": {
          "email": {
            "anyOf": [
              {
                "maxLength": 254,
                "pattern": "^[^@\\s]+@[^@\\s]+\\.[^@\\s]+$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Email"
          },
          "first_name": {
            "maxLength": 100,
            "minLength": 1,
            "title": "First Name",
            "type": "string"
          },
          "last_name": {
            "default": "",
            "maxLength": 100,
            "title": "Last Name",
            "type": "string"
          }
        },
        "required": [
          "first_name"
        ],
        "title": "PersonCreate",
        "type": "object"
      },
      "PersonOut": {
        "properties": {
          "active": {
            "title": "Active",
            "type": "boolean"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "email": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Email"
          },
          "face_count": {
            "title": "Face Count",
            "type": "integer"
          },
          "first_name": {
            "title": "First Name",
            "type": "string"
          },
          "full_name": {
            "title": "Full Name",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "last_name": {
            "title": "Last Name",
            "type": "string"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": [
          "id",
          "first_name",
          "last_name",
          "full_name",
          "email",
          "active",
          "face_count",
          "created_at",
          "updated_at"
        ],
        "title": "PersonOut",
        "type": "object"
      },
      "PersonUpdate": {
        "additionalProperties": false,
        "description": "Solo se modifican los campos enviados.",
        "properties": {
          "active": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Active"
          },
          "email": {
            "anyOf": [
              {
                "maxLength": 254,
                "pattern": "^[^@\\s]+@[^@\\s]+\\.[^@\\s]+$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Email"
          },
          "first_name": {
            "anyOf": [
              {
                "maxLength": 100,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "First Name"
          },
          "last_name": {
            "anyOf": [
              {
                "maxLength": 100,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Name"
          }
        },
        "title": "PersonUpdate",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    },
    "securitySchemes": {
      "APIKeyCookie": {
        "description": "Cookie HttpOnly que entrega POST /api/auth/login.",
        "in": "cookie",
        "name": "facetrack_session",
        "type": "apiKey"
      }
    }
  },
  "info": {
    "description": "\nReconocimiento facial, asistencia y estadísticas con **procesamiento local**.\n\n**Autenticación.** `POST /api/auth/login` entrega una cookie de sesión HttpOnly;\nel navegador la envía sola en las demás llamadas. Todas las rutas la requieren\nsalvo `GET /api/health` y el login. Las marcadas *Solo ADMIN* exigen ese rol.\n\n**Errores.** Siempre `{\"detail\": str, \"code\": str}`; los 5xx agregan\n`request_id` (también en la cabecera `X-Request-ID` de toda respuesta), que\npermite encontrar el detalle en los logs.\n\n**Privacidad.** Las imágenes se procesan en memoria y se descartan; ninguna\nrespuesta incluye embeddings.\n",
    "title": "FaceTrack API",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/personas": {
      "get": {
        "operationId": "list_persons_api_personas_get",
        "parameters": [
          {
            "in": "query",
            "name": "active",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "boolean"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Active"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/PersonOut"
                  },
                  "title": "Response List Persons Api Personas Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "APIKeyCookie": []
          }
        ],
        "summary": "List Persons",
        "tags": [
          "personas"
        ]
      },
      "post": {
        "operationId": "create_person_api_personas_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PersonCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PersonOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "APIKeyCookie": []
          }
        ],
        "summary": "Create Person",
        "tags": [
          "personas"
        ]
      }
    },
    "/api/personas/{person_id}": {
      "delete": {
        "operationId": "delete_person_api_personas__person_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "person_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Person Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "La persona no existe"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "APIKeyCookie": []
          }
        ],
        "summary": "Delete Person",
        "tags": [
          "personas"
        ]
      },
      "get": {
        "operationId": "get_person_api_personas__person_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "person_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Person Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PersonOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "La persona no existe"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "APIKeyCookie": []
          }
        ],
        "summary": "Get Person",
        "tags": [
          "personas"
        ]
      },
      "put": {
        "operationId": "update_person_api_personas__person_id__put",
        "parameters": [
          {
            "in": "path",
            "name": "person_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Person Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PersonUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PersonOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "La persona no existe"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "APIKeyCookie": []
          }
        ],
        "summary": "Update Person",
        "tags": [
          "personas"
        ]
      }
    },
    "/api/personas/{person_id}/faces": {
      "get": {
        "operationId": "list_faces_api_personas__person_id__faces_get",
        "parameters": [
          {
            "in": "path",
            "name": "person_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Person Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/FaceSampleOut"
                  },
                  "title": "Response List Faces Api Personas  Person Id  Faces Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "La persona no existe"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "APIKeyCookie": []
          }
        ],
        "summary": "List Faces",
        "tags": [
          "personas"
        ]
      },
      "post": {
        "description": "Registra una muestra facial. Solo se guarda el embedding cifrado, nunca la imagen.",
        "operationId": "enroll_face_api_personas__person_id__faces_post",
        "parameters": [
          {
            "in": "path",
            "name": "person_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Person Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_enroll_face_api_personas__person_id__faces_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FaceEnrollmentOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "Imagen inválida o corrupta"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "La persona no existe"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "El rostro no coincide con la persona o coincide con otra"
          },
          "413": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "Archivo demasiado grande"
          },
          "415": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "Formato no soportado"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "Sin rostro, varios rostros o calidad insuficiente"
          }
        },
        "security": [
          {
            "APIKeyCookie": []
          }
        ],
        "summary": "Enroll Face",
        "tags": [
          "personas"
        ]
      }
    },
    "/api/personas/{person_id}/faces/{face_id}": {
      "delete": {
        "operationId": "delete_face_api_personas__person_id__faces__face_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "person_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Person Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "face_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Face Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "La persona no existe"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "APIKeyCookie": []
          }
        ],
        "summary": "Delete Face",
        "tags": [
          "personas"
        ]
      }
    },
...
  },
...
}